+++ src/itinerary/index.tsx
@@ -76,7 +76,16 @@ const Itinerary = ({
               },
             ])}
             key={`${place.mainLabel}-${place.subLabel}-${place.isoDate}`}
+            itemProp="location"
+            itemScope
+            itemType="http://schema.org/Place"
           >
+            <meta itemProp="name" content={place.mainLabel} />
+            <meta itemProp="address" content={
+              place.subLabel && typeof place.subLabel === 'string'
+              ? place.subLabel
+              : place.mainLabel
+            } />
             <Component {...hrefProps}>
               {!small && (
                 <time dateTime={place.isoDate}>